# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for news.",
        "title": "News API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/v1",
    "paths": {
        "/files": {
            "get": {
                "description": "список файлів",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "список файлів",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUploads200"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload500"
                        }
                    }
                }
            },
            "post": {
                "description": "Створення файла",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Створення файла",
                "parameters": [
                    {
                        "description": "file upload request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FileUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload500"
                        }
                    }
                }
            }
        },
        "/files/{id}": {
            "get": {
                "description": "Інформація про файл",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Інформація про файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid файла",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload500"
                        }
                    }
                }
            },
            "delete": {
                "description": "видалення файла",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "видалення файла",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid файла",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetFileUpload500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.FileUploadDto": {
            "type": "object",
            "properties": {
                "entity_id": {
                    "type": "string",
                    "example": "0194cd77-d0ab-74db-88be-f9de341a4b5f"
                },
                "entity_type": {
                    "type": "string",
                    "example": "news_articles"
                },
                "name": {
                    "type": "string",
                    "example": "article_img_123.png"
                },
                "path": {
                    "type": "string",
                    "example": "/uploads/article_img_123.png"
                },
                "uuid": {
                    "type": "string",
                    "example": "0194cd77-d0ab-74db-88be-f9de341a4b5f"
                }
            }
        },
...
        "request.FileUploadRequest": {
            "type": "object",
            "properties": {
                "entity_id": {
                    "type": "string",
                    "example": "0194cd77-d0ab-74db-88be-f9de341a4b5f"
                },
                "entity_type": {
                    "type": "string",
                    "enum": [
                        "news_groups",
                        "news_articles"
                    ],
                    "example": "news_article"
                }
            }
        },
...
        "response.DocGetFileUpload200": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/models.FileUploadDto"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.DocGetFileUpload400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.DocGetFileUpload404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.DocGetFileUpload500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "щось пішло не так"
                    ]
                }
            }
        },
        "response.DocGetFileUploads200": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FileUploadDto"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
...
}